chore(dataobj): initial commit of value encoding
This commit introduces the dataobj package with initial utilities for
encoding and decoding individual values within a "dataset." A dataset is
the generic representation of columnar data, with a dataset.Value being
one value in one page in one column.

This initial implementation includes two encodings:

* Plain encoding (for string values), and
* delta encoding (for signed integers).

A follow-up commit will introduce bitmap encoding for efficiently
bitpacking unsigned integers.

My initial prototype of dataobj used generics rather than the
dataset.Value wrapper. However, usage of generics made it difficult to
write utilities that operates on multiple columns. While dataset.Value
is slightly less type safe, it is significantly easier to work within
the scope of a dataset.

The encoding and decoding of values is implemented to support streaming
as much as possible: individual values can be encoded and passed
immediately to a compression block. Streaming values minimizes the
number of rows that needed to be stored in memory at once on both the
write path and the read path. This constrats with the design of
parquet-go, which primarily intends for an entire page of values to be
buffered in memory prior to encoding and compression. The streaming
approach trades off slightly slower performance for memory efficiency.