chore(dataobj): add Dataset with iteration and sorting support
This commit adds dataset.Dataset, the representation of a set of columns
(each of which is split into a set of pages). A Dataset is represented
by an interface to allow for implementations where Datasets, Columns,
and Pages are backed by object storage or a cache layer.

Two utilities are introduced on top of a Dataset:

* dataset.Sort returns a new Dataset for which values are sorted in
  ascending order from a selected set of columns.

* dataset.Iter allows for iterating over rows in a dataset from a
  selected set of columns.

  This initial implementation is insufficient for reading at scale:
  there is no predicate support for filtering out pages and it lazily
  loads one page at a time during iteration. Both of these would cause
  unnecessary roundtrips to object storage.

  However, dataset.Iter is useful in the interrim for tests and in the
  implementation of dataset.Sort. It will eventually be replaced with a
  more sophisticated mechanism for scanning through a Dataset.