chore(dataobj): expose estimated size of sections
To be able to implement dataset.Builder, we must be able to estimate the
size of sections in memory.

For the logs section, this is easy: we are already building columns
gradually and can look at the cut and uncut pages to return an estimate.

However, it's more difficult for the streams section; the streams
section doesn't get built until encoding time, so we have to provide a
custom heuristic to estimate what the size would be without actually
building it.