chore(dataobj): expose building objects to public API
This commit combines the streams section and logs section into an
overall object. Appending fails when the size estimate of existing data
and the size estimate of incoming data exceeds a limit.

Basic reading support (for tests) is included. This reading support
should not be used outside of testing, as it has no way of filtering out
pages or parallelizing work.