chore(dataobj): temporarily disable sorting
During testing, @cyriltovena and I have discovered that sorting the logs
dataset is far too slow due to the number of page loads required to
iterate over the single table in sorted order.

To move testing along, sorting is being disabled for now, but it will
return once the logs section implementation is refactored to lessen the
work on the dataset.Sort implementation.