chore(dataobj): improve performance of Streams.EstimatedSize
Streams.EstimatedSize is invoked for every append to a data object
builder. This method ended up consuming 44% of total CPU time due to it
iterating over every label in every stream every call.

We can precompute the size of all labels when a new stream is added,
skipping over the iteration. This will make the method several orders of
magnitude cheaper with a large streams section.