chore(dataobj): reduce total allocations while encoding
Previously, 13% of flush time was spent growing *bytes.Buffer. The CPU
time was exacerbated by the number of calls to *bytes.Buffer.Write that
required growing the slice. Additionally, as *bytes.Buffer.Write grows
the slice exponentially, the resulting slice was much bigger than it
needed to be.

A quick benchmark showed that we can reduce the CPU time and number of
allocations by ~3x each by growing the buffer to exactly the size we
need before writing all pages.