chore(dataobj): promote metadata types to hold uint64 instead of uint32
This change promotes metadata types to be constrained to a uint64
instead of a uint32 (with the exception of CRC32, which remains uint32).

This unlocks data objects larger than 4GB and fixes an issue where the
UncompressedSize field of a column was overflowing the uint32 boundaries
when the compression ratio was high enough.