chore(dataobj): fix memory leak in dataset.Iter
Resources used by iter.Pull can be garbage collected when one of the
following are true:

1. The `stop` function returned by iter.Pull is called, or
2. The `next` function returned by iter.Pull is called past the final
   element.

dataset.Iter stored the stop function, but forgot to call it, and it
only consumed up to the final element, but not past it.

As a result, any Column passed to dataset.Iter was being permanently
retained in memory. If a MemColumn as provided, then all pages were also
unable to be garbage collected.