chore(dataobj): fix memory leak in slice truncation
dataobj reuses slices as much as possible to avoid extra allocations.
This is done by truncating a slice to length zero (`slice[:0]`).
However, any element from before the slice is truncated is still
reachable and won't be garbage collected.

Properly clearing and reusing a slice requires first calling `clear` on
the slice, and _then_ truncating the length.

There were two places in the code where clearing the slice was done
incorrectly.

To avoid running into this again, a helper package `sliceclear` has been
introduced which clears the slice and then truncates its length.