chore(dataset): allow callers to determine compression options
zstd.SpeedBestCompression offers great compression for very high CPU
cost. While the cost of this is diminished if writes are streamed as
data is read from Kafka, the CPU cost becomes untenable if compression
is done in a single pass.

Rather than picking a universal compression level for Zstd, this commit
adds CompressionOptions to allow callers to opt-in to different
compression levels based on their needs.

As a result of this, the compression for the logs and streams sections
have dropped to zstd.SpeedDefault. This decreases the compression ratio
slightly but greatly reduces CPU usage.