chore(dataobj): provide a utility for improved pooling of *bytes.Buffer
Pools of *bytes.Buffer can have their memory footprint slowly increase
as the capacity of pooled buffers grows. For example, if two callers use
a pool in parallel, all buffers in the pool will eventually have the
capacity required by the larger caller.

This problem exacerbates if a larger capacity is only needed very
rarely; usage of the pool can keep elements alive. This means the
oversized buffers can be slow to reclaim, if at all.

The new bufpool package provides a utility to bucket *bytes.Buffer pools
by the capacity of the buffers, with buckets starting from 1 KiB and
doubling in size up to 64 GiB.

Using bufpool helps bound the growth of pools and limits the lifetime of
one-off massive allocations.