chore(dataobj): use loser trees for the k-way merge
Loser trees are faster for k-way merges compared to heaps due to a lower
number of comparisons.

The memory complexity of loser trees is about twice as bad as a heap,
with a loser tree requiring k*2 elements in its slice compared to the k
elements in a heap. Despite this, the cost of a tree is still cheap
compared to the rest of the data object builder.