chore: Differentiate clients in limits package
The limits package will have two clients: a client for the
limits-frontend to connect to limits service backends, and a
client for the distributor to connect to the limits-frontend.
Naming both of these clients "client" will cause naming collisions
and also confusion when reading godoc/source code. To avoid both
of these problems, we rename the clients to IngestLimitsBackendClient
and IngestLimitsFrontendClient, and do the same for their respective
configs etc.