chore(dataobj): deduplicate decoder code across bucket and io.ReadSeeker
A previous comment identified that the code for BucketDecoder and
ReadSeekerDecoder were extremely similar, and that they could be
deduplicated by introducing some kind of "range reader" interface.

This commit inroduces such an interface, which maps perfectly to bucket
decoding. Implementations of the interface must be able to tolerate
concurrent instance of readers, which io.ReadSeeker does not. To
tolerate this while still allowing to decode data objects that are
either in-memory or backed by a file, ReadSeekerDecoder has been updated
to ReaderAtDecoder.