chore(dataobj): add high-level StreamsReader API
The StreamsReader API is a high-level API for reading streams from a
streams section. Callers can customize a StreamsReader with two types of
predicates:

1. A "LabelMatcher," which asserts both that a label exists and has a
   specific value.

2. A "LabelFilter," which asserts that a filter function returns a true
   value.

LabelMatchers are intended to be the more-performant matchers that can
utilize page filtering (based on min/max values), bloom filters, and
other types of optimizations. LabelFilters are the fallback type of
predicate.

This initial implementation is naive and missing many performance
features that the final implementation will need, lacking:

* Page filtering
* Bulk downloading pages
* Only decoding other columns after provided predicates pass