chore(dataobj): add high-level LogsReader API
The LogsReader API is a high-level API for reading logs from a logs
section. Callers can customize a LogsReader with three types of
predicates:

1. A set of stream IDs to retrieve (consumed from a sequence of IDs).

2. A "MetadataMatcher," which asserts both that a metadata key exists
   and has a specific value.

3. A "MetadataFilter," which asserts that a filter function returns
   true.

MetadataMatchers are intended to be the more-performant matchers that
can utilize page filtering (based on min/max values), bloom filters, and
other types of optimizations. MetadataFilters are the fallback type of
predicate.

Like StreamsReader, this initial implementation is naive and missing
many performance features that the final implementation will need,
lacking:

* Page filtering
* Bulk downloading pages
* Only decoding other columns after provided predicates pass