Fix(blooms): Initialize bloom gateway client only once
The bloom gateway client is used both in the bloom builder and in the
index gateways. When running Loki in SSD mode, both services are part of
the `backend` target, and therefore the client is initialised twice,
leading to duplicate metrics registration and a subsequent panic.

This commit extracts the initialisation of the bloom gateway client into
a separate service that is started once and becomes a dependency of both
bloom builder and index gateway.

Ref: #14083

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>