fix(blooms): Initialize bloom gateway client only once (#15994)
The bloom gateway client is used both in the bloom builder and in the index gateways. When running Loki in SSD mode, both services are part of the backend target, and therefore the client is initialised twice, leading to duplicate metrics registration and a subsequent panic.

This commit extracts the initialisation of the bloom gateway client into a separate service that is started once and becomes a dependency of both bloom builder and index gateway.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>
(cherry picked from commit e271e22f7c1fae676e2fce6ab002b358694d9450)