chore(dataobj): support computation of min/max values in pages and columns
This commit adds support for storing the minimum and maximum non-NULL
value encountered in a page and a column. The minimum and maximum value
of a column is determined by the minimum and maximum value across all
pages in that column.

Range statistics are opt-in to avoid bloating the statistics size when a
min/max value would be less meaningful, such as a "minimum log line" in
the message column of a logs section.

dataset.Value now implements encoding.BinaryMarshaler and
encoding.BinaryUnmarshaler to be encoded and decoded from stats, using a
simple encoding for various types:

* NULL values are encoded as an empty byte array.

* Non-NULL values are encoded first with the metadata type (uvarint),
  followed by:

  * The uvarint encoding of VALUE_TYPE_UINT64 values.
  * The varint encoding of VALUE_TYPE_INT64 values.
  * The set of bytes representing a string for VALUE_TYPE_STRING values.

Range statistics will now be computed for all column types in both
sections, with the sole exception being the message column in the logs
section.