chore(dataobj): do not panic on unrecognized compression type
Opening a page for reading should not cause a panic. An unknown
compression type may be encountered if an older version of the dataobj
reading code is trying to read an object produced by a newer version
(where new compression types are available).

Because of this, the read path should never panic when encountering an
unexpected value in an object; this was the only panic on the read path
and is now removed.