Allow provisioner to be namespaced
In the Grafana Federal Cloud we have clusters where we cannot create ClusterRole/ClusterRoleBinding
due to an increased security posture.  To ensure we can deploy the provisioner in these clusters,
this PR conditionally generates Role instead of ClusterRole if enterprise and enterprise.provisioner
are enabled and rbac.namespaced is true.

This PR also updates the provisioner job helm hooks to allow it to be customized to run on other hookTypes.
This still defaults to post-install and should have no impact to current usage.  This will allow the Grafana
Federal Cloud to use the provisioner after helm post-upgrades to attempt to create tenants as required.

Closes deployment_tools/#185454

Signed-off-by: Ryan Brady <ryan.brady@grafana.com>