chore(dataobj): improve type safety on predicates
This change adds one predicate interface per reader: StreamsPredicate
and LogsPredicate.

Predicates that can be used in both readers (such as OrPredicate) now
accept a type parameter to specify which reader they'll be used in:

    // Use as AndPredicate[StreamsPredicate] or AndPredicate[LogsPredicate]
    type AndPredicate[P Predicate] struct{ Left, Right P }

This type parameter is used to dynamically implement StreamsPredicate or
LogsPredicate.

This type parameter is only used for predicates that are valid for both
predicate types; types like MetadataFilterPredicate do not accept a type
parameter and only satisfy LogsPredicate.

This change enables checking that a predicate is valid for a reader at
compile time, removing the need for the runtime check and walkPredicate.