chore(dataobj): add rowRanges utility types
rowRanges builds on top of rowRange, where it tracks multiple ranges.

rowRanges can check for inclusion of a row, or the partial or complete
overlap with an entire rowRange. rowRanges can be intersected and
unioned together.

This will be used for row filtering: large sets of row ranges are
filtered out when pages are filtered out, while smaller sets of row
ranges are filtered out by predicates.