chore(dataobj): add columnReader for reading batches of values from a column
columnReader is the column equivalent of pageReader, and is intended to
replace iterMemColumn.

Like pageReader, columnReader supports reading values in batches and can
seek to a row in the column.

To support knowing what rows exist in which pages, a rowRange utility
type has been added. rowRange tracks an inclusive start and end row.

This commit does not remove iterMemColumn; iterMemColumn will be removed
in a later commit.