feat: Snapshopt stream overrides values on each request (#16523)
**What this PR does / why we need it**:
Caches a snapshot of resolved policies and retention temporarily to guarantee consistent results on the same request.
We evaluate retention and policies in a map when emitting some metrics so having policies or retention inconsistent along the chain for the same push request can cause us to miscalculate things or worse, interact with a nil map.