docs: Clarify thanos storage behaviour
Make the storage configuration documentation more explicit
that the Thanos object store client configuration is mututally
exclusive with the legacy configuration. When the Thanos client
is enabled, legacy configuration is all silently ignored, including
legacy named stores.

Document that the .ruler.storage configuration is only deprecated when
the thanos storage client is enabled. It is still the required, and
only, method of configuring the Ruler's storage when using the legacy
object store.

Document that .ruler.storage is mostly ignored when the thanos client is
enabled, even when only using local storage, but it must still be
present due to bug #16543. .ruler_storage is where the actual
configuration is read from in thanos mode.

Warn that the ruler storage configuration does not support named stores.