fix: Ruler does not start when using thanos client
The ruler will fail to start in "all" or "backend" startup target
with message "Ruler storage is not configured; ruler will not be
started" if Loki is configured to use the Thanos object store.

The ruler configuration should be read from the 'ruler_storage' top
level config key when thanos object store is enabled (-use-thanos-objstore=true
or storage_config.use_thanos_objstore: true in the config file).

A bug in the ruler-storage module startup check meant that the
ruler.storage block that's supposed to be ignored in thanos storage
mode was still being used to check whether the ruler was configured.
So a placeholder ruler.storage block had to be provided to get the ruler
to start, even though it would be ignored by ruler itself.

Fix, so the ruler.storage block is completely ignored and can be
entirely removed when thanos client is in use.