fix: ruler startup tries to mkdir the wrong path
When the ruler-storage module starts, it tries to mkdir the local
storage directory. This was not updated to use the path from
the ruler_storage block when the thanos object store is enabled,
and was still using the otherwise-ignored ruler.storage block instead.

Fix so it properly uses ruler_storage.filesystem.dir or
ruler_storage.local.directory when in thanos storage mode, ignoring
the non-thanos-store ruler.storage.local.directory.