fix: warn if ruler storage config is ignored
Emit a warning if ruler.storage is defined but ignored because
object_store.use_thanos_client: true causes the ruler_storage block
to be used instead. And vice versa when the ruler_storage config only
used with the thanos client is present, but the thanos client is not
enabled.

This is particularly important for ruler, because the thanos object
store client setting affects which ruler storage block is used even
when ruler is just using the local file system as a store.

The new warning looks like this:

    level=warn msg="ruler.storage exists and is not empty, but will be ignored in favour of ruler_storage because storage_config.use_thanos_objstore is true."