chore(dataobj): add utilities to window data in a dataobj
A common download size for chunked data in S3 is 8MB or 16MB. When
downloading a slice of pages, we find pages that align into an 8MB or
16MB "window" and download that entire set of pages in a single request.

This trades off fewer roundtrips for downloading garbage data: if only
two pages are downloaded, and fit within a 16MB window, the majority of
data in that 8/16MB could be outside the range of both pages.

This commit adds utilities for identifying windows. The windowing code
is made generic to permit windowing any arbitrary element in the file,
including pages and column metadata.