chore(dataobj): support page filtering for NotPredicate
Page filtering for NotPredicates require applying De Morgan's laws to
rewrite the NotPredicate to eliminate the NOT.

While evaluating predicates, flipping the result of the inner expression
is valid, but this doesn't work for page filtering: if a predicate
checks for a > 10, then pages are included if `a > 5` may be true.
However, flipping that inclusion is wrong: `!(a > 5)` is the same as
`a <= 5`, and both conditions may be true across an entire page.

By applying De Morgan's laws, we can properly check for the new
condition using page statistics to know if the page should be included;
checking whether the condition `a < 5 OR a == 5` may hold instead.