chore: Add basic building blocks of a physical plan (#16710)
This PR adds the basic components of a physical plan. For the first iteration of the new execution engine, we need to support nodes to represent limited and filter queries (with only label/line filters). These nodes are:

* DataObjScan
* SortMerge
* Projection
* Filter
* Limit

The physical plan is implemented as a [directed acyclic graph](https://en.wikipedia.org/wiki/Directed_acyclic_graph) (DAG). It has convenience functions for adding nodes (vertices) and edges, getting parents and children of a node, as well as for traversing the graph using the [depth-first-search](https://en.wikipedia.org/wiki/Depth-first_search) algorithm.

Additionally to the node types, the PR also contains the basic types of expressions, which can be properties of the nodes.

---
Signed-off-by: Christian Haudum <christian.haudum@gmail.com>