fix(helm): create namespaced RBAC when sidecar is enabled
When RBAC is namespaced, the necessary permissions for Secrets
and ConfigMaps are not granted even when the sidecar is enabled,
resulting in the permission error. In the current Helm chart logic,
the generation of RBAC for the sidecar depends on OpenShift’s
Security Context Constraints. As a result, the required RBAC is not
created in non-OpenShift environments.

To address this, we will modify the RBAC rule generation logic to
ensure that the necessary Role and RoleBinding are created
when the sidecar is enabled and RBAC is namespaced.

Signed-off-by: Tsubasa Nagasawa <toversus2357@gmail.com>