Updates metastore label/value unit tests and impl
The `forEachLabel` function has a bug in it which is filtering out labels/values incorrectly.  The `Streams` function filters all streams for just the labels we're interested in -- my previous code had an additional `if` condition for the matchers which was unnecessary.

This commit also adds more test cases for checking single and empty matchers when calling the `Labels()` and `Values()` methods.

nit: this also removes the `println` stmt from `object.go`.