chore: Move `pkg/dataobj/planner` to `pkg/engine/planner` (#16815)
The package pkg/dataobj/planner/logical was initially created within the package pkg/dataobj because of the dependency on the internal package pkg/dataobj/internal/metadatamd.

However, the package should be located inside a separate package without the dependency on the dataobj package, since the engine should be storage agnostic.

---
Signed-off-by: Christian Haudum <christian.haudum@gmail.com>