fix: bug where known streams could be rejected
This commit fixes a bug where known streams could be rejected
whenever the stream limit was exceeded. The bug occurred when
streams for a tenant where sharded over a number of partitions.
The problem was that the frontend did not take the intersection
of unknown streams from all responses, so if a stream was unknown
to the customer of partition 1, but known to the consumer of
partition 2, it would still be rejected.