chore(dataobj): immediately Close unused column readers
When calling basicReader.Reset and shrinking the number of columns, the
old columns readers will be zeroed out and available for garbage
collection. However, they are not closed, which means that they must
wait until GC runs to release any resources like Zstd decoders.

Closing them before clearing the remainder of the slice ensures that
resources get released back to the pool and are available for reuse
rather than just letting the GC clear them up.