chore(dataobj): close loser tree on defer for safety
While the loser tree guarantees that all sequences are closed if the
entire input sequence is consumed, there are cases where the entire
sequence _isn't_ consumed (like on encountering an error).

For safety, we now close the loser tree on a defer to make sure that any
resources associated with the tree are released.