chore(engine): add Instruction, Value interfaces and Return instruction
This commit starts the work to shift the plan nodes to be natively SSA.

The Instruction interface represents an SSA instruction, while the Value
interface represents some value in SSA. Some instructions are also
values.

This commit also includes the first implementation of an Instruction,
Return, which returns the final value from a plan. Return does not
implement Value.