chore: some metrics changes to ingest-limits-frontend
This commit contains some metric changes to the ingest-limits-frontend.
We will remove requests_exceeds_limits_total as it doesn't make sense
here, and if we keep it, it should be tracked in the distributor.

The second small change is removing tenant from the variable names,
we don't need to include the labels in the variable name.