feat: adds new gauge metric to track push agent latency in distributors
To assist with tracking push agent latency, we have a `mostRecentLagMs`
logged message in distributors which includes the difference of timestamps (in
millis) between when a push request is received and the most recent
timestamp of the log content being pushed.

The logged message, however, doesn't include the IP address of the
agent.  As a result, it's difficult to correlate increased latency for a
given agent.

This commit adds a new gauge metric with three labels: tenant, path (URL
endpoint from which the request came), and agent_ip_address.  This gauge
metric captures the same latency value as is logged.

If the IP address is a comma-separated list, we only take the very first
entry in the list.