feat: Add agent IPs to distributor logs
To assist with tracking push agent latency, we have a mostRecentLagMs logged message in distributors which includes the difference of timestamps (in millis) between when a push request is received and the most recent timestamp of the log content being pushed.

The logged message, however, doesn't include the IP address of the agent. As a result, it's difficult to correlate increased latency for a given agent.

This commit extracts IP addresses from the `X-Forwarded-For` HTTP header in the distributor.ParseRequest function.  The value is appended to the push request log line, along with the first agent IP address parsed from the header.