chore(engine): Integrate metastore into physical query planning (#17016)
This PR integrates the data object lookup using the metastore into the query planning.

To achieve that, I extended the `Metastore` interface with an additional function that not only returns the data object paths, but also the stream IDs of streams that match the given label matchers.

```go
type Metastore interface {
	...

	// StreamsIDs returns object store paths and stream IDs for all matching objects for the given matchers between [start,end]
	StreamIDs(ctx context.Context, start, end time.Time, matchers ...*labels.Matcher) ([]string, [][]int64, error)

	...
}
```

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>